Refactored followup to look up guild when needed
The guild object isn't available until after we build the bot,
so we cannot look up channel names in the initializer
