+++ modron/app.py
@@ -10,7 +10,8 @@ from modron.interact import attach_commands
 from modron.interact.character import HPTracker, CharacterSheet
 from modron.interact.dice_roll import DiceRollInteraction
 from modron.interact.npc import NPCGenerator
-from modron.interact.reminder import ReminderModule, FollowupModule
+from modron.interact.reminder import ReminderModule
+from modron.interact.followup import FollowupModule
 from modron.interact.stats import StatisticModule
 
 
@@ -35,7 +36,7 @@ def main(testing: bool = True):
 
     # Generate the slash command responder
     modules = [
-        FollowupModule(bot),
+        FollowupModule(),
         DiceRollInteraction(),
         HPTracker(),
         CharacterSheet(),
